In animate() function: Used fillStyle again to set the canvas to black because it gets turned to blue in my draw method and then used fillRect to essentially clear the canvas. Also, Added player.update() calls
In Player Class: Added a height and width in the constructor. Within the update method after I call the draw method, I add the velocity to the player's y axis position. Then to stop the player's from dropping through the bottom I added an ifstatement. Removed the draw() calls since we are drawing the players in the update method. 

EXPECTED BEHAVIOR AFTER COMMIT: 
Player 1 and Player 2 get drawn on the screen and fall on the y axis until they hit the bottom of the screen or canvas.height.