Added gravity
Instead of setting our velocity on the y axis to a static number I'm using a global variable called gravity that gets added to this.velocity.y constantly and essentially pulls my objects down to the bottom of the canvas. This will help especially when I start jumping the characters around the canvas.