Abstracted movement from EventListener to Animate
I'm encountering a bug in the movement where the Player's will stop moving if multiple keys are held down 

ex. Press A >> Move left with A, While Holding A Press D >> Move Right, Lift up A >> Player stops moving right. 

So I added a key monitor that instantiates all of the key objects at "pressed: false". Then added the movement to my animate function that checks if a key is pressed or not and then it will animate it.

---update---

I encountered another bug in the movement with my original if statement, I added a "last key pressed" to let the browser know what key was pressed last, we can add this check in our if statement and it will tighten up the movement controls even further.