Last Key bug fix
When I duplicated the movement for player 2, my "lastKey" was messing with my movement, the two characters could not move independent of each other. So I needed to separate them. I added a "lastKey" property to the Player class so that I can reference it individually. Also I removed the last key checks for the jump buttons.