Added Hitboxes
Added some comments to the code, as well as a hitbox property that's an object that sits on top of the player and extends out depending on how large the attack range is. For now it's just a solid 100px, also I'm drawing the hitBox out in the draw method so I can visualize it better. Also started code for collision detection, it's now working on the x axis.