Added y axis collision detection
Added y axis collision detection as well as fixed a bug that was occurring when a Player jumped, apparently the height of the object correlates with the velocity and was doing a strange bug where you would jump but only up 1 pixel, then if you pressed jump again you would jump like normal. I found that -6.3 is the lowest it could be before it would do the jump bug, so I set the y velocity to -6.5.