Attack Attack!
Added attacks to player 1 and player 2 by first adding a property called isAttacking so I can check to make sure the Player is attacking. Then I set that to true in a method called attack that can be called whenever a button is pressed. Lastly I added the check to my collision detection if statement and verified it's working.