Added an offset
Added an offset for the hitbox and detection to change the offset depending on the position of the player relative to the enemy. Probably will run into an issue with the hitbox needing to be changed to the closest Player or something, but I'll cross that bridge when I come to it.