Tonsss of stuff, see description
Added sprite images for all characters (some may or may not be used)

In classes.js - Added a bunch of notes for my code, added an attack box to go along with the original hitbox because I wanted a variable attack box due to the variation of sprite attacks. 

In main.js - added attack box to the Players, removed player 2 to focus on player 1 and enemy, added temporary keys for enemy movement testing, and COMPLETELY redid collision detection. Going to make a proper function and throw all of my helper functions into a js file in the next commit.