First working version of an Enemy AI
I have a really aggressive AI that comes directly at you and murders you almost instantly...not a bad start but I need to work on dumbing it down a bit.

In classes.js I added a class for AI, and added an aggroBox that checks for the direction of the player and if it finds one then it goes to it and attacks. 

In main.js I added a bunch of iffy code that will probably be scrapped. Just a general push for that file.